Move list repl functions to LstBase
